Fix a compilation error

+++ src/backend/x86_64/backend.cpp
@@ -317,7 +317,7 @@ void X64Backend::EmitConditionalBranch(Condition condition, Xbyak::Label& label_
       code->jmp(label_skip, Xbyak::CodeGenerator::T_NEAR);
       break;
     default:
-      fmt::print("Unsupported Condition {}\n", condition);
+      fmt::print("Unsupported Condition {}\n", (int)condition);
       std::abort();
   }
 }
